refactor: Update HeroSection and Reusable functions for better code organization and consistency
