+++ src/functions/HeroSection.ts
@@ -1,5 +1,5 @@
 import { closeHeroNavPopup } from "./ThemeTemplate";
-import { addClass, toggleClass, setSectionFocusStatus, lockFocus } from "./Reusable";
+import { addClass, removeClass, toggleClass, setSectionFocusStatus, lockFocus } from "./Reusable";
 
 const taskListSection = document.getElementById("appList") as HTMLElement;
 const heroNavPopup = document.getElementById("heroNavPopup") as HTMLElement;
@@ -30,4 +30,27 @@ export const toggleTaskListSection = (): void => {
   } else {
     setSectionFocusStatus("appList", false);
   }
+};
+
+export const toggleTaskListMenu = (): void => {
+  const taskListMenu = document.getElementById("taskMenuList") as HTMLDivElement;
+  const taskListMenuItems = document.querySelectorAll(
+    ".app__task--menu__item"
+  ) as NodeListOf<HTMLLIElement>;
+
+  toggleClass(taskListMenu, "open");
+
+  taskListMenuItems.forEach((item) => {
+    item.addEventListener("click", () => {
+      removeClass(taskListMenu, "open");
+    });
+  });
+
+  document.addEventListener("click", (event) => {
+    const target = event.target as HTMLElement;
+
+    if (target.closest("#taskMenuButton")) return;
+
+    removeClass(taskListMenu, "open");
+  });
 };
